update for tftp

+++ service/tftp/baseudp.h
@@ -8,7 +8,7 @@ public:
     BaseUdp();
     virtual ~BaseUdp(){}
 
-    virtual uint32_t write(const char* data, uint32_t size) = 0;
+    virtual uint32_t write(const char* data, size_t size) = 0;
 };
 
 #endif // BASEUDP_H
